+++ CHANGELOG.md
@@ -8,6 +8,11 @@ document.
 
 [df995e...master](https://github.com/rust-lang/rust-clippy/compare/df995e...master)
 
+### New Lints
+
+* Added [`manual_min_max`] to `complexity`
+  [#NNNNN](https://github.com/rust-lang/rust-clippy/pull/NNNNN)
+
 ## Rust 1.95
 
 Current stable, released 2026-04-16
@@ -6909,6 +6914,7 @@ Released 2018-09-13
 [`manual_map`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_map
 [`manual_memcpy`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_memcpy
 [`manual_midpoint`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_midpoint
+[`manual_min_max`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_min_max
 [`manual_next_back`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_next_back
 [`manual_non_exhaustive`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_non_exhaustive
 [`manual_noop_waker`]: https://rust-lang.github.io/rust-clippy/master/index.html#manual_noop_waker